+++ advent-2021/vector.h
@@ -1,15 +1,16 @@
 /*review*/ 
 #ifndef VECTOR
 #define VECTOR
+#include <stdint.h>
 
 typedef struct {
-    int *array;
-unsigned int size;
-    unsigned int capacity;
+    int64_t *array;
+    size_t size;
+    size_t capacity;
 } Vector;
 
 Vector *init_vector(int init_capacity);
-int get(Vector *vector, unsigned int idx);
+int get(Vector *vector, size_t idx);
 int push(Vector *vector, int value);
 int pop(Vector *vector);
 int size(Vector *vector);
